+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## v0.5.0-beta
 
+### Updates
+
+- Added more API documentation
+- Made TaskOrchestrationShim public
 - Adds `PurgeInstancesMetadataAsync` and `PurgeInstancesAsync` support and implementation to `DurableTaskGrpcClient`
 - Fix issue with mixed Newtonsoft.Json and System.Text.Json serialization.